:recycle: refactor: abstract common functions and types
